requirements.txt
Pillow
azure-cognitiveservices-vision-customvision
msrest
azure-storage-blob
azure-storage-file-datalake
pandas
opencv-python
imutils
imgaug
tensorflow
pydantic
pyyaml
